+++ src/extension.ts
@@ -7,11 +7,11 @@ import { ProjectManager } from './projectManager';
 import JobLog from './views/jobLog';
 import ProjectExplorer from './views/projectExplorer';
 import { ExtensionContext, l10n, window, workspace } from 'vscode';
-import { ProjectExplorerApi } from './projectExplorerApi';
+import { IBMiProjectExplorer } from './ibmiProjectExplorer';
 import { initialise } from './testing';
 import { ProjectFileWatcher } from './fileWatcher';
 
-export async function activate(context: ExtensionContext): Promise<ProjectExplorerApi> {
+export async function activate(context: ExtensionContext): Promise<IBMiProjectExplorer> {
 	console.log(l10n.t('Congratulations, your extension "vscode-ibmi-projectexplorer" is now active!'));
 
 	// Load Code for IBM i API