+++ src/ibmiJsonT.ts
@@ -2,6 +2,9 @@
  * (c) Copyright IBM Corp. 2023
  */
 
+/**
+ * Represents the state of project's `.ibmi.json` file.
+ */
 export interface IBMiJsonT {
     /**
      * Version
@@ -14,6 +17,9 @@ export interface IBMiJsonT {
     build?: BuildT;
 }
 
+/**
+ * Represents build options.
+ */
 export interface BuildT {
     /**
      * Objects created by building source in this directory will be put into the `objlib` library.