+++ src/views/projectExplorer/sourceFile.ts
@@ -2,15 +2,15 @@
  * (c) Copyright IBM Corp. 2023
  */
 
-import { ThemeIcon, TreeItemCollapsibleState, WorkspaceFolder, l10n } from "vscode";
+import { ThemeIcon, TreeItem, TreeItemCollapsibleState, WorkspaceFolder, l10n } from "vscode";
 import { ProjectExplorerTreeItem } from "./projectExplorerTreeItem";
-import { ContextValue } from "../../projectExplorerApi";
+import { ContextValue } from "../../ibmiProjectExplorer";
 import { SourceInfo } from "./source";
 
 /**
  * Tree item for a source file
  */
-export default class SourceFile extends ProjectExplorerTreeItem {
+export default class SourceFile extends TreeItem implements ProjectExplorerTreeItem {
   static contextValue = ContextValue.sourceFile;
   sourceInfo: SourceInfo;
 
@@ -21,7 +21,7 @@ export default class SourceFile extends ProjectExplorerTreeItem {
     this.contextValue = SourceFile.contextValue;
     this.iconPath = new ThemeIcon(`file`);
     this.tooltip = l10n.t('Name: {0}\n', sourceFileInfo.name) +
-      l10n.t('Path: {0}\n', sourceFileInfo.localUri.fsPath);
+      l10n.t('Path: {0}', sourceFileInfo.localUri.fsPath);
     this.resourceUri = sourceFileInfo.localUri;
     this.command = {
       command: `vscode.open`,