+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 All notable changes to the "vscode-ibmi-projectexplorer" extension will be documented in this file.
 
+## `1.1.0`
+
+- Add offline project support
+- Improve error handling when no connection
+
+## `1.0.2`
+
+- Rename API to IBMiProjectExplorer
+- Update docs with API changes
+
+## `1.0.1`
+
+- Update doc landing page
+- Remove types submodule, update git workflow, and rename API
+
 ## `1.0.0`
 
 - Initial release